Code smell: Line modifiers are preferred to one-line blocks.
* Replaced 6 occurances of (while .*?) *do$ with \1

  The do is unneeded in the block header form and causes problems
  with the block-to-one-line transformation.

  3 Examples:

      The code:
          while line = f.gets do
      becomes:
          while line = f.gets
      The code:
          while line = shadow.gets do
      becomes:
          while line = shadow.gets
      The code:
          while wrapper = zeros.pop do
      becomes:
          while wrapper = zeros.pop

* Replaced 15 occurances of (if .*?) *then$ with \1

  The then is unneeded in the block header form and causes problems
  with the block-to-one-line transformation.

  3 Examples:

      The code:
          if f = test_files_for(failed).find { |f| failed_trace =~ Regexp.new(f) } then
      becomes:
          if f = test_files_for(failed).find { |f| failed_trace =~ Regexp.new(f) }
      The code:
          if not Puppet[:noop] and options[:detailed_exitcodes] then
      becomes:
          if not Puppet[:noop] and options[:detailed_exitcodes]
      The code:
          if not request.post? then
      becomes:
          if not request.post?

* Replaced 765 occurances of

      ((?:if|unless|while|until) .*)
          (.*)
      end

  with

      \2 \1

  The one-line form is preferable provided:

      * The condition is not used to assign a variable
      * The body line is not already modified
      * The resulting line is not too long

  3 Examples:

      The code:
          if hdr != magic
              raise Puppet::Error, "Found #{hdr} instead of magic #{magic} in #{filename}"
          end
      becomes:
          raise Puppet::Error, "Found #{hdr} instead of magic #{magic} in #{filename}" if hdr != magic
      The code:
          if defined? @cert and @cert
              raise Puppet::Error, "Cannot replace existing certificate"
          end
      becomes:
          raise Puppet::Error, "Cannot replace existing certificate" if defined? @cert and @cert
      The code:
          if key.is_a?(String)
              key = key.intern
          end
      becomes:
          key = key.intern if key.is_a?(String)

+++ lib/puppet/defaults.rb
@@ -99,9 +99,7 @@ module Puppet
                 ENV["PATH"] = value unless value == "none"
                 paths = ENV["PATH"].split(File::PATH_SEPARATOR)
                 %w{/usr/sbin /sbin}.each do |path|
-                    unless paths.include?(path)
-                        ENV["PATH"] += File::PATH_SEPARATOR + path
-                    end
+                    ENV["PATH"] += File::PATH_SEPARATOR + path unless paths.include?(path)
                 end
                 value
             end
@@ -114,9 +112,7 @@ module Puppet
                 is in Ruby's search path",
             :call_on_define => true, # Call our hook with the default value, so we always get the libdir set.
             :hook => proc do |value|
-                if defined? @oldlibdir and $LOAD_PATH.include?(@oldlibdir)
-                    $LOAD_PATH.delete(@oldlibdir)
-                end
+                $LOAD_PATH.delete(@oldlibdir) if defined? @oldlibdir and $LOAD_PATH.include?(@oldlibdir)
                 @oldlibdir = value
                 $LOAD_PATH << value
             end
@@ -594,9 +590,7 @@ module Puppet
             :call_on_define => false,
             :desc => "(Deprecated for 'report_server') The server to which to send transaction reports.",
             :hook => proc do |value|
-                if value
-                    Puppet.settings[:report_server] = value
-                end
+                Puppet.settings[:report_server] = value if value
             end
         },
         :report_server => ["$server",
